feat: adds ampersand footer to oauth card layout - adds AmpersandFooter component - moves OauthCardLayout to its own folder - updates references to OauthCardLayout
