+++ src/components/Oauth/NoWorkspaceEntry/LandingContent.tsx
@@ -2,7 +2,7 @@ import {
   Button, FormControl, FormLabel, Heading,
 } from '@chakra-ui/react';
 
-import { OauthCardLayout } from '../OauthCardLayout';
+import { OauthCardLayout } from '../OauthCardLayout/OauthCardLayout';
 import { OAuthErrorAlert } from '../OAuthErrorAlert';
 
 type LandingContentProps = {
