+++ src/components/Oauth/OauthCardLayout/OauthCardLayout.tsx
@@ -1,5 +1,7 @@
 import { Box, Container } from '@chakra-ui/react';
 
+import { AmpersandFooter } from './AmpersandFooter';
+
 type OauthCardLayoutProps = {
   children: React.ReactNode;
 };
@@ -8,19 +10,14 @@ export function OauthCardLayout({ children }: OauthCardLayoutProps) {
   return (
     <Container>
       <Box
-        p={8}
         maxWidth="600px"
         borderWidth={1}
-        borderRadius={8}
-        boxShadow="lg"
-        textAlign={['left']}
-        margin="auto"
-        marginTop="40px"
-        bgColor="white"
-        color="gray.800"
-        fontSize="md"
+        borderRadius={4}
       >
-        {children}
+        <div style={{ padding: '3rem 2rem' }}>
+          {children}
+        </div>
+        <AmpersandFooter />
       </Box>
     </Container>
   );
