+++ src/components/Oauth/WorkspaceEntry/WorkspaceEntry.tsx
@@ -3,7 +3,7 @@ import {
   FormLabel, Heading, Input,
 } from '@chakra-ui/react';
 
-import { OauthCardLayout } from '../OauthCardLayout';
+import { OauthCardLayout } from '../OauthCardLayout/OauthCardLayout';
 import { OAuthErrorAlert } from '../OAuthErrorAlert';
 
 type WorkspaceEntryProps = {
